+++ CHANGELOG.md
@@ -6,6 +6,7 @@ A more detailed list of changes is available in the corresponding milestones for
 ### Changes to existing checks
   - **[com.google.fonts/check/metadata/undeclared_fonts]**: Clarify rationale (issue #2751)
   - **[com.google.fonts/check/metadata/filenames]**: Add rationale (issue #2751)
+  - **[com.google.fonts/check/metadata/filenames]**: Consider all files from a directory (issue #2751)
 
 
 ## 0.7.18 (2020-Feb-05)