+++ CHANGELOG.md
@@ -11,6 +11,7 @@ A more detailed list of changes is available in the corresponding milestones for
   - **[com.google.fonts/check/metadata/undeclared_fonts]**: Clarify rationale (issue #2751)
   - **[com.google.fonts/check/metadata/filenames]**: Add rationale (issue #2751)
   - **[com.google.fonts/check/metadata/filenames]**: Consider all files from a directory (issue #2751)
+  - **[com.google.fonts/check/monospace]:** Fix typo isFixedWidth is actually isFixedPitch, xAverageWidth is xAvgCharWidth
 
 
 ## 0.7.18 (2020-Feb-05)