+++ CHANGELOG.md
@@ -12,6 +12,9 @@ A more detailed list of changes is available in the corresponding milestones for
 ### Deprecated checks
   - **[com.google.fonts/check/monospace_max_advancewidth]**: (issue #2749)
 
+### Bugfixes
+  - fix generate-glyphdata command (python 3 support) (issue #2765)
+
 ### Changes to existing checks
   - **[[com.google.fonts/check/post_table_version]]**: Support CFF2 OTF Variable Fonts and add rationale (issue #2638)
   - **[[com.google.fonts/check/family/vertical_metrics]]**: Updated to check across sibling families. Check was also renamed to **com.google.fonts/check/superfamily/vertical_metrics** (issue #1487)